// client/src/Pages/Metaverse/style.ts
import styled from 'styled-components';

export const MetaverseContainer = styled.div`
  .backgroundCanvas {
    top: 0px;
    left: 0px;
    display: block;
    max-width: initial;
    max-height: initial;
    /* background-color: black; */
  }

  .characterCanvas {
    top: 0px;
    left: 0px;
    position: absolute;
  }
`;